costume - bike fix attempt

+++ include/overworld.h
@@ -32,7 +32,7 @@
 
 struct InitialPlayerAvatarState
 {
-    u8 transitionFlags;
+    u16 transitionFlags;
     u8 direction;
 };
